feat: decouple remote browser as option
