+++ packages/scaffolding-cli/templates/test/xxAMIDOxx.xxSTACKSxx.E2E.Selenium/Selenium/Factory/WebDriverFactory.cs
@@ -1,15 +1,19 @@
 ﻿using System;
-using OpenQA.Selenium;
-using OpenQA.Selenium.Chrome;
-using OpenQA.Selenium.Remote;
+using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using System.Collections.Generic;
+using OpenQA.Selenium;
+using OpenQA.Selenium.Chrome;
 
 namespace xxAMIDOxx.xxSTACKSxx.E2E.Selenium.Selenium.Factory
 {
   public class WebDriverFactory
   {
+    public static string GetNamespace()
+    {
+      return typeof(WebDriverFactory).Namespace;
+    }
+
     public static IWebDriver GetWebDriver(bool remoteBrowser = false)
     {
       // Default: chromedriver headless
@@ -23,7 +27,8 @@ namespace xxAMIDOxx.xxSTACKSxx.E2E.Selenium.Selenium.Factory
       string outPutDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
       string chromeDriverPath = Path.GetFullPath(outPutDirectory);
 
-      return remoteBrowser != false ? RemoteWebDriver() : ChromeDriver();
+      // Checks for the presence of the RemoteBrowser class. If missing, then defaults to Chrome.
+      return Type.GetType(typeName: $"{GetNamespace()}.RemoteBrowser") != null ? RemoteBrowser.RemoteWebDriver() : ChromeDriver();
     }
 
     public static IWebDriver ChromeDriver()
@@ -36,36 +41,6 @@ namespace xxAMIDOxx.xxSTACKSxx.E2E.Selenium.Selenium.Factory
       string chromeDriverPath = Path.GetFullPath(outPutDirectory);
 
       return new ChromeDriver(chromeDriverPath, caps);
-    }
-
-    [Obsolete]
-    public static IWebDriver RemoteWebDriver()
-    {
-      DesiredCapabilities caps = new DesiredCapabilities();
-
-      caps.SetCapability("user", RemoteBrowser.user);
-      caps.SetCapability("accessKey", RemoteBrowser.accessKey);
-
-      caps.SetCapability("build", "your build name");
-      caps.SetCapability("name", "your test name");
-      caps.SetCapability("platform", "Windows 10");
-      caps.SetCapability("browserName", "Chrome");
-      caps.SetCapability("version", "83.0");
-      caps.SetCapability("console", "true");
-
-      /* Note: Selenium >=3.14.0 release mars DesiredCapabilite is as obsolete.
-       * This is not yet supported by LambdaTest. Once supported, the above
-       * setup can become like below.*/
-
-      //var caps = new RemoteSessionSettings();
-      //caps.AddMetadataSetting("user", RemoteBrowser.user);
-      //caps.AddMetadataSetting("accessKey", RemoteBrowser.accessKey);
-      //caps.AddMetadataSetting("platform", "Windows 10");
-      //caps.AddMetadataSetting("browserName", "Chrome");
-      //caps.AddMetadataSetting("version", "83.0");
-      //caps.AddMetadataSetting("console", true);
-
-      return new RemoteWebDriver(RemoteBrowser.uri, caps);
     }
   }
 }
